Updated test for creating a Collaborator. - There were two tests, doing exactly the same - removed redundancy. - Updated test to follow coding standards, by adding a header, a setUp() method and easier test methods.
