+++ tests/selenium_tests/phpunit.xml
@@ -8,6 +8,7 @@
             <file>shared_tests/create_user_manager.php</file>
             <file>shared_tests/create_user_curator.php</file>
             <file>shared_tests/assign_curator_to_IVD.php</file>
+            <file>shared_tests/create_user_collaborator.php</file>
             <file>shared_tests/create_user_submitter.php</file>
             <file>shared_tests/create_disease_IVA.php</file>
             <file>shared_tests/convert_LOVD2_export_IVD.php</file>
@@ -45,7 +46,7 @@
             <file>shared_tests/install_LOVD.php</file>
             <file>shared_tests/create_user_manager.php</file>
             <file>shared_tests/create_user_curator.php</file>
-            <file>authorization_tests/create_user_collaborator.php</file>
+            <file>shared_tests/create_user_collaborator.php</file>
             <file>shared_tests/create_user_submitter.php</file>
             <file>authorization_tests/create_user_owner.php</file>
             <file>shared_tests/create_gene_IVD.php</file>
@@ -62,7 +63,7 @@
         <testsuite name="collaborator_tests">
             <file>shared_tests/install_LOVD.php</file>
             <file>shared_tests/create_gene_GJB.php</file>
-            <file>collaborator_tests/create_user_collaborator.php</file>
+            <file>shared_tests/create_user_collaborator.php</file>
             <file>shared_tests/make_user_collaborator_GJB.php</file>
             <file>collaborator_tests/login_as_collaborator.php</file>
             <file>collaborator_tests/create_user_submitter.php</file>
